Add files via upload
I just realized linear regression can't work well on our example as the y variable is categorical. So I change the research question to 'What are the factors that will affect health?', and with separate TableOnes to discuss the association between factors and genhlth/hlthpln1/sleptim1. Please take a look and complement some graphs, then we can start the slide tonight！